Changed the way error handling works for RegEdit

+++ macpac/RegEdit.cs
@@ -10,34 +10,24 @@ namespace Macpac
 		{
 			try
 			{
-				if(RegName == null)
-					Registry.LocalMachine.OpenSubKey(RegRoot, true).CreateSubKey(RegKey);
-				else
-					Registry.LocalMachine.OpenSubKey(RegRoot + RegKey, true).SetValue(RegName, RegVal, RegistryValueKind.String);
+				if(RegName == null) Registry.LocalMachine.OpenSubKey(RegRoot, true).CreateSubKey(RegKey);
+				else Registry.LocalMachine.OpenSubKey(RegRoot + RegKey, true).SetValue(RegName, RegVal, RegistryValueKind.String);
 			}
 			catch(Exception e)
 			{
-				if(e is System.Security.SecurityException || e is UnauthorizedAccessException)
-					Console.WriteLine("RegWrite Error: Permission denied.\n{0}: {1}", e.HResult, e.Message);
-				else
-					Console.WriteLine("RegWrite Error {0}: {1}", e.HResult, e.Message);
+				throw e;
 			}
 		}
 		public static void Delete(string RegKey, string RegVal = null)
 		{
 			try
 			{
-				if(RegVal == null)
-					Registry.LocalMachine.OpenSubKey(RegRoot, true).DeleteSubKeyTree(RegKey);
-				else
-					Registry.LocalMachine.OpenSubKey(RegRoot + RegKey, true).DeleteValue(RegVal);
+				if(RegVal == null) Registry.LocalMachine.OpenSubKey(RegRoot, true).DeleteSubKeyTree(RegKey);
+				else Registry.LocalMachine.OpenSubKey(RegRoot + RegKey, true).DeleteValue(RegVal);
 			}
 			catch(Exception e)
 			{
-				if(e is System.Security.SecurityException || e is UnauthorizedAccessException)
-					Console.WriteLine("RegDelete Error: Permission denied.\n{0}: {1}", e.HResult, e.Message);
-				else
-					Console.WriteLine("RegDelete Error {0}: {1}", e.HResult, e.Message);
+				throw e;
 			}
 		}
 	}
